refactor of item display
